requirements.txt
certifi==2019.11.28
python-dotenv==0.12.0
wincertstore==0.2
